// src/app/clearing/models/swconst.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class SwConst {
    sw_pkid: string;
    sw_jobid: string;
    sw_itmid: string;
    sw_year: number;
    sw_serial_no: number;
    sw_const_elementname: string;
    sw_const_code: string;
    sw_const_percent: number;
    sw_const_yieldpercent: number;
    sw_active_ingredient: string;

    rec_mode: string;
    rec_category: string;
    _globalvariables: GlobalVariables;
}